Added aerodynamics text. Resolve #6.
This is the text: Aerodynamics team is responsible for the design of the aerodynamics of our rockets. This includes the nose-cone, fuselage, control surfaces and parachutes. They are also working with rocket stabilisation using the control surfaces, trajectory predictions and parachute design. Their goal is to make sure the rocket flies straight and comes down safe.